requirements.txt
grpcio==1.6.0
grpcio-tools==1.6.0
rply==0.7.8
integer==1.0.3
requests
tarfile